Reworking core to support Serializers

+++ src/fk/core/auto_seal.py
@@ -22,7 +22,7 @@ from fk.core.workitem import Workitem
 
 
 def auto_seal(workitems: Iterable[Workitem],
-              delta: int,
+              delta: float,
               executor: Callable[[Type[AbstractStrategy], list[str], bool, datetime.datetime], None]) -> None:
     # If there are pomodoros which should have been completed X seconds ago, but are not,
     # then void them automatically.
